pyproject.toml
[project]
name = "unrealgitui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.11.0",
    "customtkinter>=5.2.2",
    "isort>=7.0.0",
    "pygithub>=2.8.1",
    "pywinpty>=3.0.2",
    "rich>=14.2.0",
    "toml>=0.10.2",
]
